+++ src/Html/HtmlBuilder.php
@@ -1,6 +1,7 @@
 <?php namespace Orchestra\Html;
 
 use BadMethodCallException;
+use Illuminate\Support\Arr;
 use Orchestra\Support\Expression;
 use Orchestra\Support\Str;
 
@@ -82,8 +83,8 @@ class HtmlBuilder extends \Illuminate\Html\HtmlBuilder
     {
         // Special consideration to class, where we need to merge both string
         // from $attributes and $defaults, then take union of both.
-        $default   = array_get($defaults, 'class', '');
-        $attribute = array_get($attributes, 'class', '');
+        $default   = Arr::get($defaults, 'class', '');
+        $attribute = Arr::get($attributes, 'class', '');
 
         $classes   = explode(' ', trim($default.' '.$attribute));
         $current   = array_unique($classes);